character moves, rotates and jumps
character moves fwd relative to the camera's direciton.

the rotation of the character changes by the camera's direciton.

the character can jump. ground checking is done in the beginning to ensure character is grounded indefinitely.

a lot of stuff happened in here. but mainly the implementation of the basic movement and ground checks were implemented.