core gameplay mechanics are implemented - folder structure rearranged
core gameplay mechanics are implemented for both leep and bownd. small bugs are present, but nothing game breaking. the implentation for the special two, which is the evasion special, is probably not implemented in the best way, but it works for now.

folder structure rearranged - not sure if the best way, but it seems all right so far.